data type changed for Distance variable

+++ airDistance.c
@@ -29,7 +29,7 @@ int main()
     lon2=lon2 * 3.14/180.0;
 
     int R=6371.0;
-    int Distance;
+    double Distance;
 
     Distance = acos(sin(lat1) * sin(lat2) + cos(lat1) * cos(lat2) * cos(lon2 - lon1)) * R;
